back/requirements.txt
Django==3.1.4
djangorestframework==3.12.2
mysqlclient==1.4.6
psycopg2==2.8.6
drf-yasg==1.20.0